Add infinite scroll helpers for common logic
