+++ apps/block_scout_web/assets/js/pages/reorgs.js
@@ -1,38 +1,22 @@
 import $ from 'jquery'
 import _ from 'lodash'
-import humps from 'humps'
 import { createStore, connectElements } from '../lib/redux_helpers.js'
-import { onScrollBottom } from '../lib/utils'
+import { withInfiniteScroll, connectInfiniteScroll } from '../lib/infinite_scroll_helpers'
 import listMorph from '../lib/list_morph'
 
 export const initialState = {
-  reorgs: [],
-
-  loadingNextPage: false,
-  pagingError: false,
-  nextPageUrl: null
+  reorgs: []
 }
 
-function reducer (state = initialState, action) {
+export const reducer = withInfiniteScroll(baseReducer)
+
+function baseReducer (state = initialState, action) {
   switch (action.type) {
     case 'ELEMENTS_LOAD': {
       return Object.assign({}, state, _.omit(action, 'type'))
     }
-    case 'LOADING_NEXT_PAGE': {
-      return Object.assign({}, state, {
-        loadingNextPage: true
-      })
-    }
-    case 'PAGING_ERROR': {
-      return Object.assign({}, state, {
-        loadingNextPage: false,
-        pagingError: true
-      })
-    }
     case 'RECEIVED_NEXT_PAGE': {
       return Object.assign({}, state, {
-        loadingNextPage: false,
-        nextPageUrl: action.msg.nextPageUrl,
         reorgs: [
           ...state.reorgs,
           ..._.map(action.msg.blocks, 'blockHtml')
@@ -57,29 +41,6 @@ const elements = {
       const newElements = state.reorgs.map((html) => $(html)[0])
       listMorph(container, newElements, { key: 'dataset.blockNumber' })
     }
-  },
-  '[data-selector="next-page-button"]': {
-    load ($el) {
-      return {
-        nextPageUrl: `${$el.hide().attr('href')}&type=JSON`
-      }
-    }
-  },
-  '[data-selector="loading-next-page"]': {
-    render ($el, state) {
-      if (state.loadingNextPage) {
-        $el.show()
-      } else {
-        $el.hide()
-      }
-    }
-  },
-  '[data-selector="paging-error-message"]': {
-    render ($el, state) {
-      if (state.pagingError) {
-        $el.show()
-      }
-    }
   }
 }
 
@@ -87,25 +48,5 @@ const $reorgListPage = $('[data-page="reorg-list"]')
 if ($reorgListPage.length) {
   const store = createStore(reducer)
   connectElements({ store, elements })
-
-  onScrollBottom(() => {
-    const { loadingNextPage, nextPageUrl, pagingError } = store.getState()
-    if (!loadingNextPage && nextPageUrl && !pagingError) {
-      store.dispatch({
-        type: 'LOADING_NEXT_PAGE'
-      })
-      $.get(nextPageUrl)
-        .done(msg => {
-          store.dispatch({
-            type: 'RECEIVED_NEXT_PAGE',
-            msg: humps.camelizeKeys(msg)
-          })
-        })
-        .fail(() => {
-          store.dispatch({
-            type: 'PAGING_ERROR'
-          })
-        })
-    }
-  })
+  connectInfiniteScroll(store)
 }
